+++ script_tools/setup_proxy.sh
@@ -18,7 +18,7 @@ get_cpu_arch () {
         return 1
     fi
 }
-get_cpu_arch "uname -p" || get_cpu_arch "uname -m" || fail get_cpu_arch "Failed to get CPU arch or unsupported CPU arch, please contact the developer."
+get_cpu_arch "uname -p" || get_cpu_arch "uname -m" || get_cpu_arch "arch" || fail get_cpu_arch "Failed to get CPU arch or unsupported CPU arch, please contact the developer."
 
 GSE_COMPARE_VERSION="1.7.2"
 NODE_TYPE=proxy
@@ -77,6 +77,9 @@ get_os_type () {
         RC_LOCAL_FILE="/etc/rc.d/rc.local"
     elif [[ "${OS_INFO,,}" =~ "hat" ]]; then
         OS_TYPE="redhat"
+        RC_LOCAL_FILE="/etc/rc.d/rc.local"
+    else
+        OS_TYPE="other"
         RC_LOCAL_FILE="/etc/rc.d/rc.local"
     fi
 }
@@ -365,10 +368,6 @@ pre_view () {
 }
 
 remove_crontab () {
-    if [ $IS_SUPER == false ]; then
-        return
-    fi
-
     local tmpcron
     tmpcron=$(mktemp "$TMP_DIR"/cron.XXXXXXX)
 
@@ -376,9 +375,11 @@ remove_crontab () {
     crontab -l | grep -v "${AGENT_SETUP_PATH}"  >"$tmpcron"
     crontab "$tmpcron" && rm -f "$tmpcron"
 
-    # 下面这段代码是为了确保修改的crontab能立即生效
-    if pgrep -x crond &>/dev/null; then
-        pkill -HUP -x crond
+    # 下面这段代码是为了确保修改的crontab立即生效
+    if [ $IS_SUPER == true ]; then
+        if pgrep -x crond &>/dev/null; then
+            pkill -HUP -x crond
+        fi
     fi
 }
 
@@ -401,6 +402,18 @@ setup_startup_scripts () {
     sed -i "\|${AGENT_SETUP_PATH}/bin/gsectl|d" $rcfile
 
     echo "[ -f $AGENT_SETUP_PATH/bin/gsectl ] && $AGENT_SETUP_PATH/bin/gsectl start >/var/log/gse_start.log 2>&1" >>$rcfile
+}
+
+
+remove_startup () {
+    if [ $IS_SUPER == false ]; then
+        return
+    fi
+
+    check_rc_file
+    local rcfile=$RC_LOCAL_FILE
+
+    sed -i "\|${AGENT_SETUP_PATH}/bin/gsectl|d" $rcfile
 }
 
 start_proxy () {
@@ -459,6 +472,16 @@ stop_proxy () {
     done
 }
 
+remove_directory () {
+    for dir in "$@"; do
+        if [ -d "$dir" ]; then
+            log remove_directory - "trying to remove directory [${dir}]"
+            rm -rf "$dir"
+            log remove_directory - "directory [${dir}] removed"
+        fi
+    done
+}
+
 remove_proxy () {
     log remove_proxy - "trying to remove proxy if exists"
     stop_proxy
@@ -468,6 +491,10 @@ remove_proxy () {
     rm -rf "${AGENT_SETUP_PATH}"
 
     if [[ "$REMOVE" = "TRUE" ]]; then
+        remove_directory ${AGENT_SETUP_PATH} ${GSE_AGENT_RUN_DIR} ${GSE_AGENT_DATA_DIR} ${GSE_AGENT_LOG_DIR}
+        remove_startup
+        log remove_proxy - "startup script removed"
+
         log remove_proxy DONE "proxy removed"
         exit 0
     else
@@ -564,6 +591,11 @@ setup_py36 () {
 }
 
 download_pkg () {
+    if [[ "${REMOVE}" == "TRUE" ]]; then
+        log download_pkg - "remove agent, no need to download package"
+        return 0
+    fi
+
     local f http_status path
 
     log download_pkg START "download gse agent package from $DOWNLOAD_URL/$PKG_NAME)."
